Controllerを画面単位に分割
