+++ src/main/java/com/todolist/webApp/User/CreateUserController.java
@@ -1,9 +1,10 @@
-package com.todolist.webApp;
+package com.todolist.webApp.User;
 
 import com.todolist.entity.UserInformation;
 import com.todolist.form.UserInformationForm;
 import com.todolist.security.SecureUserDetailsService;
 import com.todolist.service.UserInformationService;
+import com.todolist.webAppCommon.ControllerProcedure;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -21,10 +22,10 @@ import javax.validation.Valid;
  *
  * ユーザ情報変更に関する画面のControllerクラス
  * ※管理者権限で実行するものはAdminControllerにて実装
- * @link com.todolist.webApp.AdminController
+ * @link com.todolist.webApp.Admin.AdminController
  */
 @Controller
-public class UserController {
+public class CreateUserController {
 
     @Autowired
     private SecureUserDetailsService secureUserDetailsService;
@@ -41,71 +42,7 @@ public class UserController {
     }
 
     /**
-     * パスワード変更画面に遷移
-     *
-     * @param model モデル
-     * @return パスワード変更画面のアドレス
-     */
-    @RequestMapping("user/changePassword")
-    public String changePassword(Model model) {
-        controllerProcedure.addMastAttribute(model);
-        return "user/changePassword";
-    }
-
-    /**
-     * パスワードを変更し、未完了リスト画面に遷移
-     *
-     * @param oldPassword ユーザ認証のために入力された旧パスワード
-     * @param newPasswordFirst 新パスワード一回目
-     * @param newPasswordSecond 新パスワード二回目
-     * @param model モデル
-     * @return 未完了リスト画面のアドレス　※精査エラーの場合はパスワード変更画面のアドレス
-     */
-    @RequestMapping(value="user/updatePassword", method= RequestMethod.POST)
-    public String updatePassword(
-            @RequestParam() String oldPassword,
-            @RequestParam() String newPasswordFirst,
-            @RequestParam() String newPasswordSecond,
-            Model model) {
-        controllerProcedure.addMastAttribute(model);
-
-        //旧パスワードが違う場合はパスワード変更画面に戻る
-        if (!secureUserDetailsService.checkPasswordIsValidated(oldPassword)) {
-            model.addAttribute("oldPasswordMissMatch", true);
-            return "user/changePassword";
-        }
-
-        //新パスワードの一回目と二回目の入力値が異なる場合はパスワード変更画面に戻る
-        if (!newPasswordFirst.equals(newPasswordSecond)) {
-            model.addAttribute("newPasswordMissMatch", true);
-            return "user/changePassword";
-        }
-
-        userInformationService.updateUserPassword(newPasswordFirst);
-
-        return "redirect:/main/processing";
-    }
-
-    /**
-     * ユーザ新規作成画面に遷移
-     *
-     * @param model モデル
-     * @return ユーザ新規作成画面のアドレス
-     */
-    @RequestMapping("user/createUser")
-    public String createUser(Model model) {
-
-        controllerProcedure.addMastAttribute(model);
-        UserInformationForm userInformationForm = new UserInformationForm();
-
-        model.addAttribute("userInformationForm", userInformationForm);
-
-        return "user/createUser";
-    }
-
-    /**
-     * 入力されたユーザ情報に基づき、ユーザ情報を新規登録し、
-     * ログイン画面に遷移、管理者の場合は管理者メニュー画面に遷移
+     * 新規登録ボタン押下
      *
      * @param passwordFirst 登録内容：パスワード一回目
      * @param passwordSecond 登録内容：パスワード二回目
@@ -114,7 +51,7 @@ public class UserController {
      * @param model モデル
      * @return 一般ユーザ：login画面、管理者：管理者メニュー
      */
-    @RequestMapping(value="user/insertUser", method=RequestMethod.POST)
+    @RequestMapping(value="user/createUser/create", method=RequestMethod.POST)
     public String insertUser(
             @RequestParam() String passwordFirst,
             @RequestParam() String passwordSecond,
