+++ src/main/resources/templates/admin/main.html
@@ -15,8 +15,8 @@
     </div>
     <div>
         <ul>
-            <li><a th:href="@{/admin/userList}">ユーザ一覧</a></li>
-            <li><a th:href="@{/user/createUser}">ユーザ作成</a></li>
+            <li><a th:href="@{/admin/main/userList}">ユーザ一覧</a></li>
+            <li><a th:href="@{/admin/main/createUser}">ユーザ作成</a></li>
         </ul>
     </div>
 </div>
