+++ src/main/resources/templates/admin/userList.html
@@ -13,7 +13,7 @@
     <div>
         <h1>管理者メニュー</h1>
     </div>
-    <form method="post" th:action="@{/admin/editUserList}">
+    <form method="post" th:action="@{/admin/userList/editUserList}">
         <table>
             <tbody>
             <tr align="left">
