+++ src/main/resources/templates/user/changePassword.html
@@ -16,7 +16,7 @@
         </p>
     </div>
     <div>
-        <form method="post" th:action="@{/user/updatePassword}">
+        <form method="post" th:action="@{/user/changePassword/change}">
             <div class="errMsg" th:if="${oldPasswordMissMatch}">old passwordが一致しません</div>
             <div class="errMsg" th:if="${newPasswordMissMatch}">new passwordが一致しません</div>
